MFS r365891: certctl: fix unprivileged mode
The first issue was lack of quoting around INSTALLFLAGS, which set it
incorrectly and produced an error on -M.

The second issue was that we weren't actually doing the install in
unprivileged mode, making it effectively useless. This was designed to pass
through the proper metalog/unpriv flags to install(1), so just let it
happen.

Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/releng/12.2@365897 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f