Create a standalone version of sys/malloc.h
The ZSTD support for the boot loader will need to include files that
use the kernel's malloc interface. Create a standalone stub version
that's functional enough to allow this to work. There's some
limitations in this interface, and it's not quite a perfect
match. Specifically, M_WAITOK allocations can fail because there's
nothing that can be done we no memory is available.


git-svn-id: https://svn.freebsd.org/base/head@366101 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f