Add LINUX_BOOT_ABI back to arm64 GENERIC kernel.
It was removed in r355289 but forgot to return it back when new u-boot booti
support was committed.  Although booti is not the preferred method of
booting the kernel, it is very useful for the initial phase of porting
FreeBSD to a new platform or booting the kernel on various embedded boards
in an industrial environment.


git-svn-id: https://svn.freebsd.org/base/head@366196 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f