MFC r349533 (by jhb@).
Note that this does not add any functionality as IFCAP_NOMAP doesn't
exist on 12.  This is mainly an attempt to get t4_sge.c closer to what's
in head (to help with other MFCs).

r349533:
Add support for IFCAP_NOMAP to cxgbe(4).

Since cxgbe(4) uses sglist instead of bus_dma, this required updates
to the code that generates scatter/gather lists for packets.  Also,
unmapped mbufs are always sent via DMA and never as immediate data in
the payload of a work request.


git-svn-id: https://svn.freebsd.org/base/stable/12@366272 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f