Modify the NFSv4.2 VOP_COPY_FILE_RANGE() client call to return after one successful RPC.
Without this patch, the NFSv4.2 VOP_COPY_FILE_RANGE() client call would
loop until the copy "len" was completed.  The problem with doing this is
that it might take a considerable time to complete for a large "len".
By returning after a single successful Copy RPC that copied some of the
data, the application that did the copy_file_range(2) syscall will be
more responsive to signal delivery for large "len" copies.


git-svn-id: https://svn.freebsd.org/base/head@366303 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f